Restructure server route files
